CommunIT Demo
The first demo version of the new CommunIT.
As this is a work in progress things are subject to change.

Styling is definitely subject to change as well; we are currently
focused on functionality and the friendliness of the UI.

To use the new database please import the SQL Script into a NEW and
freshly created database.

Please only edit the db_class.php to connect it to your server's
database.

If there are any concerns or bugs with this version please contact the
CommunIT team.